Renamed Graphics/Lib; Removed Depend-Cycle;
