+++ src/Graphics/Library.js
@@ -1,11 +1,8 @@
-
 // This file was designed by Mike Lakner
 //
 // Hande the movement of a card from player hand to the board.
 // Card must be cloned to a graphic that is linked to a character (matching card name)
 //   and then the original card removed from the table.
-
-import { handleGraphicChange } from 'Graphics/Tokens';
 
 /**
  * Suggested and offered by Aaron to avoid an error when placing card
@@ -56,7 +53,5 @@ export function createTableGraphic(theName, theImage, theLeft, theTop, theHeight
     },
   );
 
-  // Now ensure it looks and is where we want.
-  handleGraphicChange(newObj);
   return newObj;
 }
