+++ src/Graphics/Tokens.js
@@ -2,7 +2,7 @@
 // This module was added by Mike Lakner to house the generic code supporting game tokens.
 // -----
 
-import { createTableGraphic } from 'Graphics/Lib';
+import { createTableGraphic } from 'Graphics/Library';
 import { mbHandleLoopMove } from 'Graphics/Loop';
 import { mbHandleReticleMove } from 'Graphics/Reticle';
