+++ src/modules/Board.js
@@ -1,7 +1,7 @@
 // -----
 // This module was added by Mike Lakner to house the code supporting what is happening on the game board.
 // -----
-import { createTableGraphic } from 'Graphics/Lib';
+import { createTableGraphic } from 'Graphics/Library';
 
 /**
  * @param {number} theLeft
